Rotation related corrections
- There are some conditions deleted while placing rotation angles code, I added them.

- Rotation lock was screwed up since CM 12. Fixed it by fetching allowed rotations from CM's
allowed rotations setting.

- Also, a CAF commit had killed rotation lock ability.

Change-Id: I8f1b468249c68e7b6514d1a96bdb3fc638af84fd
Signed-off-by: eray orçunus <erayorcunus@gmail.com>

Window rotation: fix up sensor logic

The previous logic was incorrect because it would always use the
lastRotation when the activity requested SCREEN_ORENTATION_FULL_SENSOR
or SCREEN_ORENTATION_FULL_USER, but it should have used the
sensorRotation in that case.

Adjust the logic to be closer to the AOSP implementation to use the
sensorRotation when the activity has set so, or the user has
specifically expressed they want their device to rotate to that
orientation

Ref: CYNGNOS-1163

Change-Id: Ic0e38fdc97b03cda3fa0f85c0f2d327f66a23f6a
Signed-off-by: Roman Birg <roman@cyngn.com>