OMS: integrate OverlayManagerService into framework
Hand over ownership of overlays to OverlayManagerService.

Changes to a package's overlays are propagated via the configuration
change mechanism. This provides a well-defined point to modify an
application's assets while the application is paused.

The system maintains a global Configuration object. Since this object
is copied to applications, it has to be light-weight and must not
disclose package private information. Add a new field
Configuration.assetSeq. When the OverlayManagerService makes any change
to any package's set of overlays, the global assetSeq is incremented.
This change is visible to an application during its next
onConfigurationChanged. To mitigate this broadcast, each Resources
object keeps a hash of its loaded resources. If the
OverlayManagerService change doesn't concern the application, the hash
stays the same and the application quickly aborts the no-op update.

Add a corresponding package name to each Resources object. This is
required to apply the correct overlays for special cases, such as for
Resources obtained via PackageManager.getResourcesForApplication.

Consolidate how overlays targeting the system and overlays targeting
regular applications are handled. Previously, system overlays were
handled as a special case. Now, everything is handled identically. As a
side effect, the call to idmap --scan during Zygote boot has become
obsolete and is removed.

Deprecate and remove use of SCAN_TRUSTED_OVERLAY from
PackageManagerService. Previously, the flag was used to restrict what
paths overlays were allowed to be installed in. Now, overlay packages
are first class packages and it is up to the OverlayManagerService to
decide whether to use the overlay or not.

Deprecate ApplicationInfo.resourcesDirs. The old constant value is
obsolete now that the set of overlays can change dynamically. Query the
OverlayManagerService instead.

Co-authored-by: Martin Wallgren <martin.wallgren@sonymobile.com>
Signed-off-by: Zoran Jovanovic <zoran.jovanovic@sonymobile.com>
Change-Id: Idc96dae6fc075d5373aa055bbf50e919136d7353

Conflicts:
	cmds/idmap/scan.cpp
	core/java/android/app/ActivityThread.java
	core/java/android/app/ApplicationThreadNative.java
	core/java/android/app/IApplicationThread.java
	libs/androidfw/AssetManager.cpp
	services/core/java/com/android/server/SystemServiceManager.java